Fix MEDIUM and LOW issues: list_split, str_from_utf8, sized loads, deterministic RC
MEDIUM fixes:
- #13: Implement list_split_first and list_split_last lowlevel ops
- #14: Fix str_from_utf8 to use proper validation via builtins.str.fromUtf8
- #15: Replace fixed 32-element arg array with Scratch(ValueLocation) in callLambdaBodyDirect
- #17: Widen enum_dispatch.tag from u8 to u16 to match num_functions
- #18: Add errdefer to LirExprStore/MonoExprStore initCapacity
- #19: Fix list_get over-read by using correctly-sized loads for 1/2/4-byte elements
- #20: Fix wildcard placeholder in TailRecursion to use layout-aware zero

LOW fixes:
- #21: Make RC op ordering deterministic by sorting HashMap keys before emission
- #22: Add debug assertions for parallel span lengths (join params, proc args)
- #23: Add debug guards against .none sentinel IDs in store accessors
- #24: Add debug assertion in registerSymbolDef against duplicate definitions
- #25: Remove dead LambdaSet and LambdaSetRepresentation types

+++ src/mono/MonoIR.zig
@@ -137,7 +137,7 @@ pub const ClosureRepresentation = union(enum) {
         /// Number of functions in the set
         num_functions: u16,
         /// This function's tag value (position in lambda set)
-        tag: u8,
+        tag: u16,
         /// All members of the lambda set (for dispatch at call sites)
         lambda_set: LambdaSetMemberSpan,
     },
